+++ bot/management/commands/send_penny_chat_template.py
@@ -8,11 +8,11 @@ import slack
 
 from bot.processors.pennychat import (
     shared_message_preview_template,
-    shared_message_template,
-    organizer_edit_after_share_template,
 )
+from bot.tasks import organizer_edit_after_share_template, shared_message_template
 from pennychat.models import PennyChatInvitation
 from users.models import get_or_create_user_profile_from_slack_id
+# TODO! make ticket for getting this working again
 
 
 class Command(BaseCommand):
