+++ bot/processors/pennychat.py
@@ -2,9 +2,9 @@ import json
 from datetime import datetime, timedelta
 import logging
 from pytz import timezone, utc
-import urllib.parse
 import requests
 
+from bot.tasks import post_organizer_edit_after_share_template, shared_message_template
 from bot.utils import chat_postEphemeral_with_fallback
 from pennychat.models import (
     PennyChat,
@@ -17,7 +17,6 @@ from bot.processors.filters import (
     is_block_interaction_event,
     has_event_type, has_callback_id)
 from users.models import (
-    get_or_create_user_profile_from_slack_ids,
     get_or_create_user_profile_from_slack_id,
     UserProfile,
 )
@@ -51,158 +50,6 @@ def get_time_options():
         for dt in datetime_range(datetime(2019, 1, 1, 0), datetime(2019, 1, 2, 0), timedelta(minutes=15))
     ]
     return time_options
-
-
-def datetime_template(penny_chat):
-    timestamp = int(penny_chat.date.astimezone(utc).timestamp())
-    date_text = f'<!date^{timestamp}^{{date_pretty}} at {{time}}|{penny_chat.date}>'
-    return date_text
-
-
-def shared_message_template(penny_chat, user_name, include_rsvp=False):
-    start_date = penny_chat.date.astimezone(utc).strftime('%Y%m%dT%H%M%SZ')
-    end_date = (penny_chat.date.astimezone(utc) + timedelta(hours=1)).strftime('%Y%m%dT%H%M%SZ')
-    google_cal_url = 'https://calendar.google.com/calendar/render?' \
-                     'action=TEMPLATE&text=' \
-        f'{urllib.parse.quote(penny_chat.title)}&dates=' \
-        f'{start_date}/{end_date}&details=' \
-        f'{urllib.parse.quote(penny_chat.description)}'
-
-    body = [
-        {
-            'type': 'section',
-            'text': {
-                'type': 'mrkdwn',
-                'text': f'_*{user_name}* invited you to a new Penny Chat!_'
-            }
-        },
-        {
-            'type': 'section',
-            'text': {
-                'type': 'mrkdwn',
-                'text': f'*Title*\n{penny_chat.title}'
-            }
-        },
-        {
-            'type': 'section',
-            'text': {
-                'type': 'mrkdwn',
-                'text': f'*Description*\n{penny_chat.description}'
-            }
-        },
-        {
-            'type': 'section',
-            'text': {
-                'type': 'mrkdwn',
-                'text': f'*Date and Time*\n{datetime_template(penny_chat)}'
-            },
-            'accessory': {
-                'type': 'button',
-                'text': {
-                    'type': 'plain_text',
-                    'text': 'Add to Google Calendar :calendar:',
-                    'emoji': True
-                },
-                'url': google_cal_url
-            }
-        },
-    ]
-
-    if include_rsvp:
-        body.append(
-            {
-                'type': 'actions',
-                'elements': [
-                    {
-                        'type': 'button',
-                        'text': {
-                            'type': 'plain_text',
-                            'text': 'Count me in :thumbsup:',
-                            'emoji': True,
-                        },
-                        'action_id': PENNY_CHAT_CAN_ATTEND,
-                        'value': json.dumps({PENNY_CHAT_ID: penny_chat.id}),  # TODO should this be a helper function?
-                        'style': 'primary',
-                    },
-                    {
-                        'type': 'button',
-                        'text': {
-                            'type': 'plain_text',
-                            'text': 'I can\'t make it :thumbsdown:',
-                            'emoji': True,
-                        },
-                        'action_id': PENNY_CHAT_CAN_NOT_ATTEND,
-                        'value': json.dumps({PENNY_CHAT_ID: penny_chat.id}),
-                        'style': 'primary',
-                    }
-
-                ]
-            }
-        )
-
-    return body
-
-
-def organizer_edit_after_share_template(slack_client, penny_chat_invitation):
-    shares = []
-    users = get_or_create_user_profile_from_slack_ids(
-        comma_split(penny_chat_invitation.invitees),
-        slack_client=slack_client,
-    )
-    for slack_user_id in comma_split(penny_chat_invitation.invitees):
-        shares.append(users[slack_user_id].real_name)
-
-    organizer = get_or_create_user_profile_from_slack_ids(
-        [penny_chat_invitation.organizer_slack_id],
-        slack_client=slack_client,
-    ).get(penny_chat_invitation.organizer_slack_id)
-
-    if len(penny_chat_invitation.channels) > 0:
-        for channel in comma_split(penny_chat_invitation.channels):
-            shares.append(f'<#{channel}>')
-
-    if len(shares) == 1:
-        share_string = shares[0]
-    elif len(shares) == 2:
-        share_string = ' and '.join(shares)
-    elif len(shares) > 2:
-        shares[-1] = f'and {shares[-1]}'
-        share_string = ', '.join(shares)
-
-    shared_message_preview_blocks = shared_message_template(penny_chat_invitation, organizer.real_name) + [
-        {
-            'type': 'divider'
-        },
-        {
-            'type': 'section',
-            'text': {
-                'type': 'mrkdwn',
-                'text': f'*:point_up: You just shared this invitation with:* {share_string}. '
-                    'We will notify you as invitees respond.\n\n'
-                    'In the meantime if you need to update the event, click the button below.'
-            }
-        },
-        {
-            'type': 'actions',
-            'elements': [
-                {
-                    'type': 'button',
-                    'text': {
-                        'type': 'plain_text',
-                        'text': 'Edit Details :pencil2:',
-                        'emoji': True,
-                    },
-                    # TODO should this be a helper function?
-                    'value': json.dumps({PENNY_CHAT_ID: penny_chat_invitation.id}),
-                    'action_id': PENNY_CHAT_EDIT,
-                    'style': 'primary',
-                }
-
-            ]
-        },
-    ]
-
-    return shared_message_preview_blocks
 
 
 def penny_chat_details_modal(penny_chat_invitation):
@@ -440,15 +287,15 @@ class PennyChatBotModule(BotModule):
     @has_event_type([VIEW_SUBMISSION, VIEW_CLOSED])
     @has_callback_id(PENNY_CHAT_DETAILS)
     def submit_details_and_share(self, event):
+        # TODO! remove organizer from invitee list
         view = event['view']
+        # TODO! hide penny chat id in the value and get rid of fetch by view
         penny_chat_invitation = PennyChatInvitation.objects.get(view=view['id'])
         state = view['state']['values']
 
         if event['type'] == VIEW_CLOSED:
-            self.slack_client.chat_postMessage(
-                channel=penny_chat_invitation.organizer_slack_id,
-                blocks=organizer_edit_after_share_template(self.slack_client, penny_chat_invitation),
-            )
+            # TODO make NOW
+            post_organizer_edit_after_share_template(view['id'])
             return
 
         if len(penny_chat_invitation.invitees.strip()) == 0 and len(penny_chat_invitation.channels.strip()) == 0:
@@ -506,11 +353,8 @@ class PennyChatBotModule(BotModule):
         penny_chat_invitation.shares = json.dumps(shares)
         penny_chat_invitation.save()
 
-        self.slack_client.chat_postMessage(
-            channel=penny_chat_invitation.organizer_slack_id,
-            blocks=organizer_edit_after_share_template(self.slack_client, penny_chat_invitation),
-        )
-
+        # TODO make NOW
+        post_organizer_edit_after_share_template(view['id'])
         return
 
     @is_block_interaction_event
@@ -618,6 +462,7 @@ class PennyChatBotModule(BotModule):
             logging.exception('error in penny chat attendance selection')
 
 
+# TODO! do we still need this?
 def comma_split(comma_delimited_string):
     """normal string split for  ''.split(',') returns [''], so using this instead"""
     return [x for x in comma_delimited_string.split(',') if x]
