--- penny_university/management/commands/trash.py
@@ -1,8 +1,0 @@
-import time
-
-from django.core.management.base import BaseCommand
-
-
-class Command(BaseCommand):
-    def handle(self, *args, **options):
-        print('YYYYYYYYY')
